# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: "Ethiopian Cuisine Products and Recipes Store API"
  version: "1.0.0"
  description: "This API allows users to explore and purchase authentic Ethiopian cuisine products and discover traditional recipes. The API provides routes to manage products, users, and shopping carts, ensuring a seamless shopping and culinary experience for those interested in Ethiopian culture and cuisine."
servers:
  - url: "http://localhost:3005/api"
    description: Development server
  - url: "https://te-enat-shop-api.onrender.com/api"
    description: Production server
paths:
...
  /products/{id}:
    get:
      summary: Get a product by ID
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          description: The product ID
          schema:
            type: string
      operationId: getProduct
      responses:
        200:
          description: A single product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        404:
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a product by ID
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          description: The product ID
          schema:
            type: string
      responses:
        204:
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        404:
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update a product by ID
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          description: The product ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        200:
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        404:
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /recipes/{id}:
    get:
      summary: Get a recipe by ID
      tags:
        - Recipes
      parameters:
        - in: path
          name: id
          required: true
          description: The recipe ID
          schema:
            type: string
      operationId: getRecipe
      responses:
        200:
          description: A single recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        404:
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a recipe by ID
      tags:
        - Recipes
      parameters:
        - in: path
          name: id
          required: true
          description: The recipe ID
          schema:
            type: string
      responses:
        204:
          description: Recipe deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        404:
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update a recipe by ID
      tags:
        - Recipes
      parameters:
        - in: path
          name: id
          required: true
          description: The recipe ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Recipe"
      responses:
        200:
          description: Recipe updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        404:
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /packages/{id}:
    get:
      summary: Get a package by ID
      tags:
        - Packages
      parameters:
        - in: path
          name: id
          required: true
          description: The package ID
          schema:
            type: string
      operationId: getPackage
      responses:
        200:
          description: A single package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        404:
          description: Package not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a package by ID
      tags:
        - Packages
      parameters:
        - in: path
          name: id
          required: true
          description: The package ID
          schema:
            type: string
      responses:
        204:
          description: Package deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        404:
          description: Package not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update a package by ID
      tags:
        - Packages
      parameters:
        - in: path
          name: id
          required: true
          description: The package ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Package"
      responses:
        200:
          description: Package updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        404:
          description: Package not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    Product:
      type: object
      required:
        - _id
        - name
        - shortDescription
        - pricing
        - image
        - categories
        - features
      properties:
        _id:
          type: string
          description: The unique identifier for the product.
          example: "60d21b4667d0d8992e610c85"
        name:
          type: string
          description: The name of the product.
          example: "Awesome Gadget"
        shortDescription:
          type: string
          description: A brief description of the product.
          example: "A multi-functional gadget that makes your life easier."
        pricing:
          type: array
          description: An array of pricing and sizes options for the product.
          items:
            type: object
            properties:
              size:
                type: ProductSize
                description: The weight option for the product.
              price:
                type: number
                format: float
                description: The price of the product for the given size.
                example: 19.99
            required:
              - size
              - price
        image:
          type: string
          description: A URL link for the product Image.
          format: uri
          example: "https://example.com/product-image.jpg"
        categories:
          type: array
          description: A list of category names the product belongs to.
          items:
            $ref: "#/components/schemas/Category"
        subCategories:
          type: array
          description: A list of sub-category names the product belongs to.
          items:
            $ref: "#/components/schemas/SubCategory"
        features:
          type: object
          description: An optional list of benefits associated with the product.
          properties:
            _id:
              type: string
              description: The unique identifier for the feature.
            value:
              type: array
              items:
                type: Feature
          required:
            - id
            - value
        totalSales:
          type: integer
          description: An optional field for the total number of sales.
          example: 5000
...
    Package:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        image:
          type: string
        price:
          type: integer
          format: float
        cookingTime:
          type: integer
          format: int
        ingredientsQuantity:
          type: integer
          format: int
        peoplesQuantity:
          type: integer
          format: int
      required:
        - _id
        - name
        - image
        - price
        - cookingTime
        - ingredientsQuantity
        - peoplesQuantity
...
    Review:
      type: object
      properties:
        reviewerName:
          type: string
          description: ID of the user who wrote the review
        rating:
          type: integer
          format: int32
          description: Rating given to the recipe (e.g., 1 to 5)
          minimum: 1
          maximum: 5
        comment:
          type: string
          description: Comment provided by the user for the recipe
        createdAt:
          type: Date
          format: date-time
          description: The date and time when the review was created
      required:
        - userId
        - recipeId
        - rating
        - comment
      example:
        userId: "647e59e1f9a4e7a1b4c3d7f4"
        recipeId: "647e59f1f9a4e7a1b4c3d7f5"
        rating: 5
        comment: "Amazing recipe! The instructions were easy to follow, and the dish turned out great."
        createdAt: "2023-08-20T12:45:00Z"
...
    Recipe:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the recipe
        name:
          type: string
          description: Name of the recipe
        description:
          type: string
          description: A brief description of what the recipe is about
        image:
          type: string
          description: URL to an image of the recipe
        ingredients:
          type: array
          items:
            type: Ingredient
        instructions:
          type: array
          items:
            $ref: "#/components/schemas/Instruction"
        servings:
          type: integer
          description: Number of servings the recipe can be prepared for
        prepTime:
          type: string
          description: Preparation time for the recipe (e.g., "30 minutes")
        difficulty:
          type: string
          enum: ["Easy", "Medium", "Hard"]
          description: Difficulty level of the recipe
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
            description: Names of categories the recipe belongs to
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - _id
        - name
        - description
        - image
        - ingredients
        - instructions
        - servings
        - prepTime
        - difficulty
        - createdAt
        - categories
    Category:
      type: object
      properties:
        nameInHebrew:
          type: string
          description: "The name of the product category in Hebrew"
          example: "תבלינים"
        nameInEnglish:
          type: string
          description: "The value of the product category in English"
          example: "Spices"
      required:
        - nameInHebrew
        - nameInEnglish
    SubCategory:
      type: object
      properties:
        nameInHebrew:
          type: string
        nameInEnglish:
          type: string
        nameOfParentCategory:
          type: string
      required:
        - nameInHebrew
        - nameInEnglish
        - nameOfParentCategory
...
    Instruction:
      type: object
      properties:
        step:
          type: integer
        description:
          type: string
        _id:
          type: string
      required:
        - step
        - description
...
    RandomItem:
      oneOf:
        - $ref: "#/components/schemas/Product"
        - $ref: "#/components/schemas/Recipe"
        - $ref: "#/components/schemas/Package"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          oneOf:
            - type: object
              $ref: "#/components/schemas/RandomItem"
            - type: array
              items:
                $ref: "#/components/schemas/RandomItem"
      required:
        - success
        - message
        - data
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message